+++ project/frontend/src/components/ProfileTabContent.tsx
@@ -1,0 +1,287 @@
+import { motion } from 'framer-motion';
+import { User, Zap, Heart, Compass, Loader2, Sparkles, Instagram } from 'lucide-react';
+import { useState } from 'react';
+
+import { buildTasteProfileSections, type SavedItem } from '../App';
+
+type ProfileTabContentProps = {
+  items: SavedItem[];
+  onGoToFeed: () => void;
+  onSelectItem: (item: SavedItem) => void;
+  onTasteChange: React.Dispatch<React.SetStateAction<string>>;
+  taste: string;
+  user: { id: number; username: string } | null;
+};
+
+export function ProfileTabContent({
+  items,
+  onGoToFeed,
+  onSelectItem,
+  onTasteChange,
+  taste,
+  user,
+}: ProfileTabContentProps) {
+  const tasteSections = buildTasteProfileSections(taste);
+  const [isGeneratingTaste, setIsGeneratingTaste] = useState(false);
+  const [isSharingProfile, setIsSharingProfile] = useState(false);
+
+  const handleGenerateTaste = async () => {
+    setIsGeneratingTaste(true);
+    try {
+      const res = await fetch('/api/generate-taste', { method: 'POST' });
+      const data = await res.json();
+
+      if (data.success) {
+        onTasteChange(data.summary);
+      } else {
+        alert(data.message || "취향 분석에 실패했습니다.");
+      }
+    } catch (error) {
+      console.error("Taste generation failed:", error);
+      alert("취향 분석 중 에러가 발생했습니다.");
+    } finally {
+      setIsGeneratingTaste(false);
+    }
+  };
+
+  const handleShareProfile = async () => {
+    if (!taste) {
+      alert('먼저 취향 프로필을 생성해 주세요.');
+      return;
+    }
+
+    const shareText = `My POSE! 취향 프로필\n\n닉네임: ${user?.username || 'Anonymous'}\n\n${taste}\n\n#POSE #취향프로필 #Aesthetic`;
+    const instagramUrl = /iPhone|iPad|iPod|Android/i.test(navigator.userAgent)
+      ? 'instagram://camera'
+      : 'https://www.instagram.com/';
+
+    setIsSharingProfile(true);
+    try {
+      if (navigator.clipboard?.writeText) {
+        await navigator.clipboard.writeText(shareText);
+      } else {
+        const textarea = document.createElement('textarea');
+        textarea.value = shareText;
+        textarea.setAttribute('readonly', 'true');
+        textarea.style.position = 'absolute';
+        textarea.style.left = '-9999px';
+        document.body.appendChild(textarea);
+        textarea.select();
+        document.execCommand('copy');
+        document.body.removeChild(textarea);
+      }
+
+      if (navigator.share) {
+        try {
+          await navigator.share({
+            title: '내 취향 프로필',
+            text: shareText,
+          });
+          return;
+        } catch (error) {
+          if (error instanceof DOMException && error.name === 'AbortError') {
+            alert('공유가 취소되었어요. 복사된 텍스트를 인스타그램에 붙여넣어 공유할 수 있어요.');
+            return;
+          }
+          console.error('Native share failed, falling back to Instagram helper:', error);
+        }
+      }
+
+      window.open(instagramUrl, '_blank', 'noopener,noreferrer');
+      alert('취향 프로필이 클립보드에 복사되었어요. 인스타그램에서 붙여넣어 공유해 보세요.');
+    } catch (error) {
+      console.error('Profile share failed:', error);
+      alert('프로필 텍스트 복사에 실패했습니다. 브라우저 권한을 확인해 주세요.');
+    } finally {
+      setIsSharingProfile(false);
+    }
+  };
+
+  return (
+    <motion.div
+      key="profile"
+      initial={{ opacity: 0, y: 20 }}
+      animate={{ opacity: 1, y: 0 }}
+      exit={{ opacity: 0, y: -20 }}
+      className="max-w-4xl mx-auto space-y-16 py-8"
+    >
+      <div className="flex flex-col md:flex-row items-start md:items-end justify-between gap-8 border-b border-black/10 pb-12">
+        <div className="space-y-5">
+          <div className="w-28 h-28 rounded-full bg-gradient-to-tr from-blue-600 via-yellow-300 to-purple-600 p-[3px] shadow-2xl">
+            <div className="w-full h-full bg-white rounded-full flex items-center justify-center overflow-hidden">
+              <User className="w-12 h-12 text-black" />
+            </div>
+          </div>
+          <div className="space-y-1">
+            <h2 className="text-5xl font-black tracking-tighter text-black uppercase">
+              {user?.username || 'Anonymous'}
+            </h2>
+            <p className="text-xs font-black text-gray-400 uppercase tracking-widest bg-gray-100 inline-block px-3 py-1 rounded-full">
+              POSE Creator No. {user?.id || '000'}
+            </p>
+          </div>
+        </div>
+        <div className="text-left md:text-right max-w-xs">
+          <p className="text-2xl font-black text-black leading-tight tracking-tighter uppercase bg-clip-text text-transparent bg-gradient-to-r from-black to-gray-500">
+            "My Vibe is<br/>my POSE."
+          </p>
+        </div>
+      </div>
+
+      <div className="grid grid-cols-1 lg:grid-cols-12 gap-12">
+        <div className="lg:col-span-4 space-y-6">
+          <div className="flex items-center gap-3 text-[10px] font-black uppercase tracking-[0.3em] text-blue-600">
+            <Zap className="w-3.5 h-3.5" fill="currentColor" />
+            Taste Analysis
+          </div>
+          <h3 className="text-3xl font-black tracking-tighter leading-none uppercase">
+            The Patterns<br/>of your<br/>Choices.
+          </h3>
+          <div className="h-2 w-12 bg-black rounded-full" />
+        </div>
+
+        <div className="lg:col-span-8">
+          {taste ? (
+            <div className="relative overflow-hidden rounded-[3rem] border border-black/5 bg-white shadow-[0_30px_80px_-40px_rgba(0,0,0,0.35)]">
+              <div className="absolute inset-0 bg-[radial-gradient(circle_at_top_left,_rgba(168,85,247,0.18),_transparent_30%),radial-gradient(circle_at_top_right,_rgba(59,130,246,0.12),_transparent_28%),linear-gradient(135deg,_rgba(250,250,250,0.96),_rgba(244,244,245,0.98))]" />
+              <div className="relative p-8 md:p-10 space-y-8">
+                <div className="flex flex-col gap-5 md:flex-row md:items-start md:justify-between">
+                  <div className="space-y-3">
+                    <span className="inline-flex items-center gap-2 rounded-full border border-black/10 bg-white/80 px-4 py-2 text-[10px] font-black uppercase tracking-[0.3em] text-gray-500 backdrop-blur">
+                      <Heart className="h-3.5 w-3.5 text-pink-500" fill="currentColor" />
+                      Signature Taste
+                    </span>
+                    <div>
+                      <h4 className="text-3xl md:text-4xl font-black tracking-tighter text-black">
+                        {user?.username || 'Anonymous'}님의 취향 한 장 요약
+                      </h4>
+                      <p className="mt-2 text-sm md:text-base font-medium text-gray-500">
+                        AI가 모아본 취향의 결, 좋아하는 무드, 다시 찾게 될 영감 포인트예요.
+                      </p>
+                    </div>
+                  </div>
+                  <div className="rounded-[2rem] bg-black px-5 py-4 text-white shadow-xl shadow-black/10">
+                    <p className="text-[10px] font-black uppercase tracking-[0.25em] text-white/60">Pose Index</p>
+                    <p className="mt-2 text-3xl font-black tracking-tight">{Math.max(items.length, tasteSections.length).toString().padStart(2, '0')}</p>
+                    <p className="mt-1 text-xs font-medium text-white/70">Collected inspirations analyzed</p>
+                  </div>
+                </div>
+
+                <div className="grid gap-4 md:grid-cols-2">
+                  {tasteSections.map((section, index) => (
+                    <div
+                      key={`${section.title}-${index}`}
+                      className={`rounded-[2rem] border bg-gradient-to-br ${section.accent} p-6 backdrop-blur-sm`}
+                    >
+                      <div className="flex items-start justify-between gap-3">
+                        <div>
+                          <p className="text-[10px] font-black uppercase tracking-[0.3em] text-gray-400">Taste Cue {String(index + 1).padStart(2, '0')}</p>
+                          <h5 className="mt-2 text-xl font-black tracking-tight text-black capitalize">{section.title}</h5>
+                        </div>
+                        <Compass className="h-5 w-5 text-gray-400" />
+                      </div>
+                      <div className="mt-5 space-y-3">
+                        {section.body.map((line, lineIndex) => (
+                          <div key={`${section.title}-${lineIndex}`} className="flex gap-3 rounded-2xl bg-white/80 px-4 py-3 shadow-sm shadow-black/5">
+                            <div className="mt-1 h-2.5 w-2.5 shrink-0 rounded-full bg-black" />
+                            <p className="text-sm font-medium leading-6 text-gray-700">{line}</p>
+                          </div>
+                        ))}
+                      </div>
+                    </div>
+                  ))}
+                </div>
+
+                <div className="rounded-[2rem] border border-dashed border-black/10 bg-white/70 p-5 text-sm font-medium leading-7 text-gray-600">
+                  <span className="mr-2 inline-flex rounded-full bg-black px-3 py-1 text-[10px] font-black uppercase tracking-[0.25em] text-white">Share Tip</span>
+                  공유 버튼을 누르면 프로필 문구를 먼저 클립보드에 복사한 뒤, 인스타그램으로 이동할 수 있게 도와드려요.
+                </div>
+
+                <div className="flex flex-wrap items-center justify-end gap-3 border-t border-gray-200/80 pt-6">
+                  <button
+                    onClick={handleGenerateTaste}
+                    disabled={isGeneratingTaste}
+                    className="flex items-center gap-2 rounded-2xl bg-white px-6 py-3 text-xs font-black uppercase tracking-widest text-black shadow-sm transition-all hover:bg-gray-100"
+                  >
+                    {isGeneratingTaste ? <Loader2 className="w-4 h-4 animate-spin" /> : <Sparkles className="w-4 h-4" />}
+                    {isGeneratingTaste ? 'Analyzing...' : 'Re-Analyze'}
+                  </button>
+                  <button
+                    onClick={handleShareProfile}
+                    disabled={isSharingProfile}
+                    className="flex items-center gap-2 rounded-2xl bg-gradient-to-r from-purple-600 to-blue-600 px-6 py-3 text-xs font-black uppercase tracking-widest text-white shadow-lg shadow-purple-500/30 transition-all hover:from-purple-700 hover:to-blue-700"
+                  >
+                    {isSharingProfile ? <Loader2 className="w-4 h-4 animate-spin" /> : <Instagram className="w-4 h-4" />}
+                    {isSharingProfile ? 'Preparing...' : 'Share to IG'}
+                  </button>
+                </div>
+              </div>
+            </div>
+          ) : (
+            <div className="py-24 bg-gray-50 rounded-[3rem] flex flex-col items-center justify-center text-center space-y-8 px-4 border-2 border-dashed border-gray-200">
+              <div className="w-20 h-20 bg-white rounded-full flex items-center justify-center shadow-sm">
+                <Sparkles className="w-8 h-8 text-yellow-400" fill="currentColor" />
+              </div>
+
+              {items.length > 0 ? (
+                <div className="space-y-6 flex flex-col items-center">
+                  <p className="text-gray-500 font-bold text-lg max-w-sm">
+                    충분한 영감이 모였습니다.<br />당신의 무의식적인 패턴을 꺼내볼까요?
+                  </p>
+                  <button
+                    onClick={handleGenerateTaste}
+                    disabled={isGeneratingTaste}
+                    className="px-10 py-4 bg-black text-white rounded-full text-sm font-black tracking-widest uppercase hover:bg-gray-800 hover:scale-105 active:scale-95 transition-all flex items-center gap-3 disabled:opacity-50 shadow-xl"
+                  >
+                    {isGeneratingTaste ? <Loader2 className="w-5 h-5 animate-spin" /> : <Zap className="w-5 h-5 text-yellow-400" fill="currentColor" />}
+                    {isGeneratingTaste ? 'Analyzing POSE...' : 'Analyze My POSE'}
+                  </button>
+                </div>
+              ) : (
+                <div className="space-y-6 flex flex-col items-center">
+                  <p className="text-gray-400 font-bold text-lg">아직 수집된 영감이 없습니다.</p>
+                  <button
+                    type="button"
+                    onClick={onGoToFeed}
+                    className="px-8 py-3 bg-white border-2 border-gray-200 text-black rounded-full text-xs font-black uppercase tracking-widest hover:border-black transition-all"
+                  >
+                    Explore Feed
+                  </button>
+                </div>
+              )}
+            </div>
+          )}
+        </div>
+      </div>
+
+      <div className="space-y-8 pt-12">
+        <div className="flex items-center justify-between border-b border-black/5 pb-6">
+          <h4 className="text-xl font-black uppercase tracking-tighter text-black">
+            Recent Inspirations
+          </h4>
+          <div className="text-xs font-bold text-gray-400 uppercase tracking-widest bg-gray-100 px-3 py-1 rounded-full">
+            {items.length} Items
+          </div>
+        </div>
+        <div className="grid grid-cols-3 md:grid-cols-4 lg:grid-cols-6 gap-3">
+          {items.slice(0, 12).map((item) => (
+            <div
+              key={item.id}
+              className="aspect-square bg-gray-100 overflow-hidden cursor-pointer group relative rounded-2xl"
+              onClick={() => onSelectItem(item)}
+            >
+              <img
+                src={item.image_url?.startsWith('http') ? item.image_url : item.image_url ? `/api/images/${item.image_url}` : 'https://via.placeholder.com/400x500?text=No+Image'}
+                className="w-full h-full object-cover grayscale opacity-80 group-hover:opacity-100 group-hover:grayscale-0 transition-all duration-500 group-hover:scale-110"
+                referrerPolicy="no-referrer"
+                onError={(e) => {
+                  (e.target as HTMLImageElement).src = 'https://via.placeholder.com/400x500?text=No+Image';
+                }}
+              />
+            </div>
+          ))}
+        </div>
+      </div>
+    </motion.div>
+  );
+}
